Added three new tests (#85)
I added three tests, one to test for a 1 input, one to test for a large numbered input, such as 1500, and one to test for a negative number input. I also removed an extra instance of getList which ended up doubling our ending list size for the computation results. This should result in more accurate tests for our program